requirements.txt
# ============================================================================
# Visitor Management System - Python Dependencies
# Python 3.11 Compatible - For Render & AWS Lambda Deployment
# ============================================================================

# ============================================================================
# Core FastAPI and ASGI Dependencies
# ============================================================================
fastapi==0.121.3
starlette==0.49.3
uvicorn==0.34.0
mangum==0.20.0

# ============================================================================
# AWS Services (S3, Lambda)
# ============================================================================
boto3==1.35.71
botocore==1.35.99
s3transfer==0.10.4
jmespath==1.0.1

# ============================================================================
# Database
# ============================================================================
sqlalchemy==2.0.44
psycopg2-binary==2.9.10
greenlet==3.2.4

# ============================================================================
# Authentication & Security
# ============================================================================
bcrypt==5.0.0
python-jose[cryptography]==3.5.0
cryptography==46.0.3

# ============================================================================
# Data Validation (Pydantic Stack)
# ============================================================================
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.6.1
email-validator>=2.3.0

# ============================================================================
# HTTP & Networking
# ============================================================================
anyio==4.11.0
idna==3.11
sniffio==1.3.1
python-multipart==0.0.20
h11==0.16.0

# ============================================================================
# SMS & Communication (Twilio)
# ============================================================================
twilio==9.3.0

# ============================================================================
# QR Code Generation
# ============================================================================
qrcode[pil]==7.4.2
Pillow==11.0.0
pypng==0.20220715.0
colorama==0.4.6

# ============================================================================
# Type Hints & Annotations
# ============================================================================
typing-extensions==4.15.0
annotated-types==0.7.0
annotated-doc==0.0.4
typing-inspection==0.4.2

# ============================================================================
# Cryptography Dependencies
# ============================================================================
cffi==2.0.0
pycparser==2.23
ecdsa==0.19.1
pyasn1==0.6.1
rsa==4.9.1

# ============================================================================
# Utilities
# ============================================================================
six==1.17.0
python-dateutil==2.9.0.post0
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
requests==2.32.5
click==8.3.1
python-dotenv==1.2.1
PyJWT==2.10.1

# ============================================================================
# Twilio Dependencies (auto-installed with twilio, but pinned for stability)
# ============================================================================
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiohappyeyeballs==2.6.1
attrs==25.4.0
frozenlist==1.8.0
multidict==6.7.0
propcache==0.4.1
yarl==1.22.0

# ============================================================================
# DNS (required by email-validator)
# ============================================================================
dnspython==2.8.0